+++ source/states/DebugState.c
@@ -112,7 +112,7 @@ static bool DebugState_processMessage(DebugState this __attribute__ ((unused)), 
 	{
 		case kKeyPressed:
 			{
-				u16 pressedKey = *((u16*)Telegram_getExtraInfo(telegram));
+				u32 pressedKey = *((u32*)Telegram_getExtraInfo(telegram));
 
 				if(pressedKey & K_LL)
 				{
